feat: [EXT-3415] move all package names to @contentful
Because the package names we were using were never claimed on
npm, there is a danger that a malicious actor could upload
something to one of the names. Then someone who is new to the
ecosystem might have tried to install one of these packages.

This PR adds `@contentful/` in front of all package names. This
means that no one outside of the Contentful org on NPM is able to
claim any of these names.